// src/shared/enums/login-event.enum.ts
export enum LOGIN_EVENT {
    IS_USERNAME_EXIST = 'IS_USERNAME_EXIST',
    GET_IP_ADDRESS = 'GET_IP_ADDRESS',
    USER_ADD = 'USER_ADD',
    FORCE_LOGOUT = 'FORCE_LOGOUT',
    IS_NOREF_EXIST = 'IS_NOREF_EXIST',
    LOGIN_NOTIF = 'LOGIN_NOTIF',
    NOTIF_PESAN = 'NOTIF_PESAN',

    FILE_NOTIF = 'FILE_NOTIF',
    FILE_UPLOAD = 'FILE_UPLOAD',
    ENGINE_FILE_NOTIF = 'ENGINE_FILE_NOTIF',
    USER_ADD_MULTIPLE = 'USER_ADD_MULTIPLE',

    ENGINE_FILE_NOTIF2 = 'ENGINE_FILE_NOTIF2',
    FILE_NOTIF2 = 'FILE_NOTIF2',
}